Create a frequency histogram for codes

Create a frequency histogram for codes

code_freq_hist(
  x,
  codes = ".*",
  sortByFreq = "decreasing",
  forceRootStripping = FALSE,
  trimSourceIdentifiers = 20,
  ggplot2Theme = ggplot2::theme(legend.position = "bottom"),
  silent = rock::opts$get("silent")
)

Arguments

x
A parsed source(s) object.

codes
A regular expression to select codes to include.

sortByFreq
Whether to sort by frequency decreasingly (decreasing, the default), increasingly (increasing), or alphabetically (NULL).

forceRootStripping
Force the stripping of roots, even if they are different.

trimSourceIdentifiers
If not NULL, the number of character to trim the source identifiers to.

ggplot2Theme
Can be used to specify theme elements for the plot.

silent
Whether to be chatty or silent.

Value

a ggplot2::ggplot().