Create a coding scheme

This function can be used to specify a coding scheme that can then be used in analysis.

create_codingScheme(
  id,
  label,
  codes,
  codingInstructions = NULL,
  description = "",
  source = ""
)

codingScheme_peterson

codingScheme_levine

codingScheme_willis

Arguments

id
An identifier for this coding scheme, consisting only of letters, numbers, and underscores (and not starting with a number).

label
A short human-readable label for the coding scheme.

codes
A character vector with the codes in this scheme.

codingInstructions
Coding instructions; a named character vector, where each element is a code's coding instruction, and each element's name is the corresponding code.

description
A description of this coding scheme (i.e. for information that does not fit in the label).

source
Optionally, a description, reference, or URL of a source for this coding scheme.

Format

An object of class rock_codingScheme of length 5.

An object of class rock_codingScheme of length 5.

An object of class rock_codingScheme of length 5.

Value

The coding scheme object.

Details

A number of coding schemes for cognitive interviews are provided:

codingScheme_peterson

Coding scheme from Peterson, Peterson & Powell, 2017

codingScheme_levine

Coding scheme from Levine, Fowler & Brown, 2005

codingScheme_willis

Coding scheme from Willis, 1999