Expand categorical attribute variables to a series of dichotomous variables

Expand categorical attribute variables to a series of dichotomous variables

expand_attributes(
  data,
  attributes,
  valueLabels = NULL,
  prefix = "",
  glue = "__",
  suffix = "",
  falseValue = 0,
  trueValue = 1,
  valueFirst = TRUE,
  append = TRUE
)

Arguments

data
The data frame, normally the $mergedSources data frame that exists in the object returned by a call to parse_sources().

attributes
The name of the attribute(s) to expand.

valueLabels
It's possible to use different names for the created variables than the values of the attributes. This can be set with the valueLabels argument. If only one attribute is specified, pass a named vector for valueLabels, and if multiple attributes are specified, pass a named list of named vectors, where the name of each vector corresponds to an attribute passed in attributes. The names of the vector elements must correspond to the values of the attributes (see the example).

prefix, suffix
The prefix and suffix to add to the variables names that are returned.

glue
The glue to paste the first part ad the second part of the composite variable name together.

falseValue, trueValue
The values to set for rows that, respectively, do not match and do match an attribute value.

valueFirst
Whether to insert the attribute value first, or the attribute name, in the composite variable names.

append
Whether to append the columns to the supplied data frame or not.

Value

A data.frame

Examples

### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-1.rock");

### Parse single example source
parsedExample <- rock::parse_source(exampleFile);

### Create a categorical attribute column
parsedExample$mergedSourceDf$age_group <-
  rep(c("<18", "18-30", "31-60", ">60"), each=13);

### Expand to four logical columns
parsedExample$mergedSourceDf <-
  rock::expand_attributes(
    parsedExample$mergedSourceDf,
    "age_group",
    valueLabels =
      c(
        "<18" = "youngest",
        "18-30" = "youngish",
        "31-60" = "oldish",
        ">60" = "oldest"
       ),
    valueFirst = FALSE
);

### Show result
table(parsedExample$mergedSourceDf$age_group,
      parsedExample$mergedSourceDf$age_group__youngest);
#>        
#>          0  1
#>   <18    0 13
#>   >60   13  0
#>   18-30 13  0
#>   31-60 13  0
table(parsedExample$mergedSourceDf$age_group,
      parsedExample$mergedSourceDf$age_group__oldish);
#>        
#>          0  1
#>   <18   13  0
#>   >60   13  0
#>   18-30 13  0
#>   31-60  0 13