Generic underlying recoding function

This function contains the general set of actions that are always used when recoding a source (e.g. check the input, document the justification, etc). Users should normally never call this function.

generic_recoding(
  input,
  codes,
  func,
  filter = TRUE,
  output = NULL,
  outputPrefix = "",
  outputSuffix = "_recoded",
  decisionLabel = NULL,
  justification = NULL,
  justificationFile = NULL,
  preventOverwriting = rock::opts$get("preventOverwriting"),
  encoding = rock::opts$get("encoding"),
  silent = rock::opts$get("silent"),
  ...
)

Arguments

input
One of 1) a character string specifying the path to a file with a source; 2) an object with a loaded source as produced by a call to load_source(); 3) a character string specifying the path to a directory containing one or more sources; 4) or an object with a list of loaded sources as produced by a call to load_sources().

codes
The codes to process

func
The function to apply.

filter
Optionally, a filter to apply to specify a subset of the source(s) to process (see get_source_filter()).

output
If specified, the coded source will be written here.

outputPrefix, outputSuffix
The prefix and suffix to add to the filenames when writing the processed files to disk, in case multiple sources are passed as input.

decisionLabel
A description of the (recoding) decision that was taken.

justification
The justification for this action.

justificationFile
If specified, the justification is appended to this file. If not, it is saved to the justifier::workspace(). This can then be saved or displayed at the end of the R Markdown file or R script using justifier::save_workspace().

preventOverwriting
Whether to prevent overwriting existing files when writing the files to output.

encoding
The encoding to use.

silent
Whether to be chatty or quiet.

...
Other arguments to pass to fnc.

Value

Invisibly, the recoded source(s) or source(s) object.