Read sources from a directory, parse them, and show coded fragments and code tree

This function combines successive calls to parse_sources(), collect_coded_fragments() and show_inductive_code_tree().

inspect_coded_sources(
  path,
  parse_args = list(extension = "rock|dct", regex = NULL, recursive = TRUE,
    ignoreOddDelimiters = FALSE, encoding = rock::opts$get("encoding"), silent =
    rock::opts$get("silent")),
  fragments_args = list(codes = ".*", context = 0),
  inductive_tree_args = list(codes = ".*", output = "both", headingLevel = 3),
  deductive_tree_args = list()
)

Arguments

path
The path containing the sources to parse and inspect.

parse_args
The arguments to pass to parse_sources().

fragments_args
The arguments to pass to collect_coded_fragments().

inductive_tree_args
The arguments to pass to show_inductive_code_tree().

deductive_tree_args
Not yet implemented.

Value

The parsedSources object.

Examples

### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Inspect sources
rock::inspect_coded_sources(examplePath);

#>                      levelName
#> 1  codes                      
#> 2   ¦--parentCode1            
#> 3   ¦   ¦--childCode1         
#> 4   ¦   ¦   ¦--grandchildCode1
#> 5   ¦   ¦   ¦--grandchildCode2
#> 6   ¦   ¦   °--grandchildCode3
#> 7   ¦   ¦--childCode2         
#> 8   ¦   °--childCode3         
#> 9   ¦--parentCode2            
#> 10  ¦   ¦--childCode4         
#> 11  ¦   ¦   ¦--grandchildCode4
#> 12  ¦   ¦   ¦--grandchildCode5
#> 13  ¦   ¦   °--grandchildCode6
#> 14  ¦   °--childCode5         
#> 15  ¦       °--grandchildCode7
#> 16  ¦--someOtherCode          
#> 17  ¦--att_ins_eval           
#> 18  ¦--chairs                 
#> 19  ¦--inductFather           
#> 20  ¦   ¦--inducChild3        
#> 21  ¦   ¦--inducChild4        
#> 22  ¦   °--inducChild5        
#> 23  ¦--inductMother           
#> 24  ¦   ¦--inducChild1        
#> 25  ¦   °--inducChild2        
#> 26  ¦--internet               
#> 27  ¦--oaken_chests           
#> 28  ¦--people                 
#> 29  ¦--tables                 
#> 30  ¦--Topic1                 
#> 31  ¦--Topic2                 
#> 32  ¦--code1                  
#> 33  ¦--code2                  
#> 34  ¦   °--code4              
#> 35  ¦--code3                  
#> 36  ¦--code5                  
#> 37  ¦--bla                    
#> 38  ¦   ¦--subbla             
#> 39  ¦   ¦   ¦--bla2           
#> 40  ¦   ¦   °--bla3           
#> 41  ¦   °--subbla2            
#> 42  ¦--flups                  
#> 43  ¦--pets                   
#> 44  ¦   °--cats               
#> 45  ¦       °--siamese        
#> 46  ¦--property               
#> 47  ¦   °--hairy              
#> 48  ¦--relationships          
#> 49  ¦   °--love               
#>              levelName
#> 1 codes               
#> 2  ¦--content_adequacy
#> 3  ¦--retrieval       
#> 4  °--understanding   
#>                      levelName
#> 1  codes                      
#> 2   ¦--parentCode1            
#> 3   ¦   ¦--childCode1         
#> 4   ¦   ¦   ¦--grandchildCode1
#> 5   ¦   ¦   ¦--grandchildCode2
#> 6   ¦   ¦   °--grandchildCode3
#> 7   ¦   ¦--childCode2         
#> 8   ¦   °--childCode3         
#> 9   ¦--parentCode2            
#> 10  ¦   ¦--childCode4         
#> 11  ¦   ¦   ¦--grandchildCode4
#> 12  ¦   ¦   ¦--grandchildCode5
#> 13  ¦   ¦   °--grandchildCode6
#> 14  ¦   °--childCode5         
#> 15  ¦       °--grandchildCode7
#> 16  ¦--someOtherCode          
#> 17  ¦--att_ins_eval           
#> 18  ¦--chairs                 
#> 19  ¦--inductFather           
#> 20  ¦   ¦--inducChild3        
#> 21  ¦   ¦--inducChild4        
#> 22  ¦   °--inducChild5        
#> 23  ¦--inductMother           
#> 24  ¦   ¦--inducChild1        
#> 25  ¦   °--inducChild2        
#> 26  ¦--internet               
#> 27  ¦--oaken_chests           
#> 28  ¦--people                 
#> 29  ¦--tables                 
#> 30  ¦--Topic1                 
#> 31  ¦--Topic2                 
#> 32  ¦--code1                  
#> 33  ¦--code2                  
#> 34  ¦   °--code4              
#> 35  ¦--code3                  
#> 36  ¦--code5                  
#> 37  ¦--bla                    
#> 38  ¦   ¦--subbla             
#> 39  ¦   ¦   ¦--bla2           
#> 40  ¦   ¦   °--bla3           
#> 41  ¦   °--subbla2            
#> 42  ¦--flups                  
#> 43  ¦--pets                   
#> 44  ¦   °--cats               
#> 45  ¦       °--siamese        
#> 46  ¦--property               
#> 47  ¦   °--hairy              
#> 48  ¦--relationships          
#> 49  ¦   °--love               
#>              levelName
#> 1 codes               
#> 2  ¦--content_adequacy
#> 3  ¦--retrieval       
#> 4  °--understanding   

#> Parsed 14 sources, with filenames 'attitude.dct', 'attributes.rock', 'behavior.dct', 'ci_example_1.rock', 'example-1.rock', 'intention.dct', 'longer-test.rock', 'lorum-ipsum.rock', 'lorum-ipsum2.rock', 'merging-test-1-primary.rock', 'merging-test-1-secondary.rock', 'perceived_behavioral_control.dct', 'perceived_norm.dct' & 'second-test-file.rock'.