Merge source files by different coders

This function takes sets of sources and merges them using the utterance identifiers (UIDs) to match them.

merge_sources(input, output, outputPrefix = "",
  outputSuffix = "_merged", primarySourcesRegex = ".*",
  primarySourcesIgnoreRegex = outputSuffix, primarySourcesPath = input,
  coderId = "\\[\\[coderId=([a-zA-Z0-9._-]+)\\]\\]",
  idForOmittedCoderIds = "noCoderId", codeRegexes = c(codes =
  "\\[\\[([a-zA-Z0-9._>-]+)\\]\\]"), idRegexes = c(caseId =
  "\\[\\[cid=([a-zA-Z0-9._-]+)\\]\\]", stanzaId =
  "\\[\\[sid=([a-zA-Z0-9._-]+)\\]\\]", coderId =
  "\\[\\[coderId=([a-zA-Z0-9._-]+)\\]\\]"),
  sectionRegexes = c(paragraphs = "---paragraph-break---", secondary =
  "---<[a-zA-Z0-9]?>---"),
  uidRegex = "\\[\\[uid=([a-zA-Z0-9._-]+)\\]\\]",
  autoGenerateIds = c("stanzaId"), persistentIds = c("caseId",
  "coderId"), noCodes = "^uid:|^uid=|^dct:|^ci:", recursive = TRUE,
  primarySourcesRecursive = recursive, filenameRegex = ".*",
  delimiterRegEx = "^---$", ignoreRegex = "^#", overwrite = FALSE,
  ignoreOddDelimiters = FALSE, postponeDeductiveTreeBuilding = TRUE,
  encoding = "UTF-8", silent = TRUE, inheritSilence = FALSE)

Arguments

input
The directory containing the input sources.

output
The path to the directory where to store the merged sources. This path will be created with a warning if it does not exist. An exception is if "same" is specified - in that case, every file will be written to the same directory it was read from.

outputPrefix, outputSuffix
A pre- and/or suffix to add to the filename when writing the merged sources (especially useful when writing them to the same directory).

primarySourcesRegex
A regular expression that specifies how to recognize the primary sources (i.e. the files used as the basis, to which the codes from other sources are added).

primarySourcesIgnoreRegex
A regular expression that specifies which files to ignore as primary files.

primarySourcesPath
The path containing the primary sources.

coderId
A regular expression specifying the coder identifier, specified similarly to the codeRegexes.

idForOmittedCoderIds
The identifier to use for utterances that do not have a coder id (i.e. utterance that occur in a source that does not specify a coder id, or above the line where a coder id is specified).

codeRegexes
These are named character vectors with one or more regular expressions. For codeRegexes, these specify how to extract the codes (that were used to code the sources). For idRegexes, these specify how to extract the different types of identifiers. For sectionRegexes, these specify how to extract the different types of sections. The codeRegexes and idRegexes must each contain one capturing group to capture the codes and identifiers, respectively.

idRegexes
These are named character vectors with one or more regular expressions. For codeRegexes, these specify how to extract the codes (that were used to code the sources). For idRegexes, these specify how to extract the different types of identifiers. For sectionRegexes, these specify how to extract the different types of sections. The codeRegexes and idRegexes must each contain one capturing group to capture the codes and identifiers, respectively.

sectionRegexes
These are named character vectors with one or more regular expressions. For codeRegexes, these specify how to extract the codes (that were used to code the sources). For idRegexes, these specify how to extract the different types of identifiers. For sectionRegexes, these specify how to extract the different types of sections. The codeRegexes and idRegexes must each contain one capturing group to capture the codes and identifiers, respectively.

autoGenerateIds
The names of the idRegexes that, if missing, should receive autogenerated identifiers (which consist of 'autogenerated_' followed by an incrementing number).

persistentIds
The names of the idRegexes for the identifiers which, once attached to an utterance, should be attached to all following utterances as well (until a new identifier with the same name is encountered, after which that identifier will be attached to all following utterances, etc).

noCodes
This regular expression is matched with all codes after they have been extracted using the codeRegexes regular expression (i.e. they're matched against the codes themselves without, for example, the square brackets in the default code regex). Any codes matching this noCodes regular expression will be ignored, i.e., removed from the list of codes.

recursive, primarySourcesRecursive
Whether to read files from sub-directories (TRUE) or not.

filenameRegex
Only files matching this regular expression are read.

delimiterRegEx
The regular expression that is used to extract the YAML fragments.

ignoreRegex
The regular expression that is used to delete lines before any other processing. This can be used to enable adding comments to sources, which are then ignored during analysis.

overwrite
Whether to overwrite existing files or not.

ignoreOddDelimiters
If an odd number of YAML delimiters is encountered, whether this should result in an error (FALSE) or just be silently ignored (TRUE).

postponeDeductiveTreeBuilding
Whether to imediately try to build the deductive tree(s) based on the information in this file (FALSE) or whether to skip that. Skipping this is useful if the full tree information is distributed over multiple files (in which case you should probably call parse_sources instead of parse_source).

encoding
The encoding of the file to read (in file).

silent
Whether to provide (FALSE) or suppress (TRUE) more detailed progress updates.

inheritSilence
If not silent, whether to let functions called by merge_sources inherit that setting.

Value

Invisibly, a list of the parsed, primary, and merged sources.