Parse a source

This function parses a source and the contained identifiers, sections, and codes.

parse_source(x, codeRegexes = c(code =
  "\\[\\[([a-zA-Z0-9._>-]+)\\]\\]"), idRegexes = c(caseId =
  "\\[\\[cid=([a-zA-Z0-9._-]+)\\]\\]", stanzaId =
  "\\[\\[sid=([a-zA-Z0-9._-]+)\\]\\]"), sectionRegexes = c(paragraphs
  = "---paragraph-break---", secondary = "---<[a-zA-Z0-9]?>---"),
  autoGenerateIds = c("stanzaId"), persistentIds = c("caseId"),
  inductiveCodingHierarchyMarker = ">", delimiterRegEx = "^---$",
  ignoreRegex = "^#", ignoreOddDelimiters = FALSE, silent = FALSE)

# S3 method for rockParsedSource
print(x, prefix = "### ", ...)

Arguments

x
The character vector containing the source, for example as obtained through [base::readLines()].

codeRegexes, idRegexes, sectionRegexes
These are named character vectors with one or more regular expressions. For `codeRegexes`, these specify how to extract the codes (that were used to code the sources). For `idRegexes`, these specify how to extract the different types of identifiers. For `sectionRegexes`, these specify how to extract the different types of sections. The `codeRegexes` and `idRegexes` must each contain one capturing group to capture the codes and identifiers, respectively.

autoGenerateIds
The names of the `idRegexes` that, if missing, should receive autogenerated identifiers (which consist of 'autogenerated_' followed by an incrementing number).

persistentIds
The names of the `idRegexes` for the identifiers which, once attached to an utterance, should be attached to all following utterances as well (until a new identifier with the same name is encountered, after which that identifier will be attached to all following utterances, etc).

inductiveCodingHierarchyMarker
For inductive coding, this marker is used to indicate hierarchical relationships between codes. The code at the left hand side of this marker will be considered the parent code of the code on the right hand side. More than two levels can be specified in one code (for example, if the `inductiveCodingHierarchyMarker` is '>', the code `grandparent>child>grandchild` would indicate codes at three levels.

delimiterRegEx
The regular expression that is used to extract the YAML fragments.

ignoreRegex
The regular expression that is used to delete lines before any other processing. This can be used to enable adding comments to sources, which are then ignored during analysis.

ignoreOddDelimiters
If an odd number of YAML delimiters is encountered, whether this should result in an error (`FALSE`) or just be silently ignored (`TRUE`).

silent
Whether to provide or suppress more detailed progress updates.