Extract DCT specifications from all files in a directory

This function processes all DCT specifications in a directory.

parse_sources(path, extension = "rock|dct", regex, codeRegexes = c(code
  = "\\[\\[([a-zA-Z0-9._>-]+)\\]\\]"), idRegexes = c(caseId =
  "\\[\\[cid=([a-zA-Z0-9._-]+)\\]\\]", stanzaId =
  "\\[\\[sid=([a-zA-Z0-9._-]+)\\]\\]"),
  autoGenerateIds = c("stanzaId"), sectionRegexes = c(paragraphs =
  "---paragraph-break---", secondary = "---<[a-zA-Z0-9]?>---"),
  inductiveCodingHierarchyMarker = ">", delimiterRegEx = "^---$",
  metadataContainers = c("metadata"), codesContainers = c("codes",
  "dct"), ignoreRegex = "^#", ignoreOddDelimiters = FALSE,
  encoding = "UTF-8", silent = TRUE)

# S3 method for rockParsedSources
print(x, prefix = "### ", ...)

# S3 method for rockParsedSources
plot(x, ...)

Arguments

path
The path from where to read the DCT files.

extension
The extension of the file; convenient alternative to specifying a `regex`.

regex
A regular expression; when provided, overrides the `extension` argument to guide file selection in the `path`.

delimiterRegEx
The regular expression specifying how the YAML fragments specifying the constructs are delimited. Should normally never be changed.

metadataContainers, codesContainers
A character vector listing the containers of the metadata and code specifications in the YAML fragments. Because only specifications are read that are stored in these containers, the files can contain YAML fragments with other data, too, without interfering with the parsing of the ROCK specifications.

ignoreOddDelimiters
Whether to ignore a final odd delimiter, if encountered.

silent
Whether to provide information on progress.

Value

An object of class `dctRawSpecListSet` for processing by [parse_dct_specs()].

Details

This function is called by [process_dir()]; it is normally not necessary to call this function directly.

Examples

# NOT RUN {
extract_dct_dir("A:/path/to/some/directory");
# }