Remove one or more codes

These functions remove one or more codes from a source, and make it easy to justify that decision.

recode_delete(
  input,
  codes,
  filter = TRUE,
  output = NULL,
  childrenReplaceParents = TRUE,
  recursiveDeletion = FALSE,
  decisionLabel = NULL,
  justification = NULL,
  justificationFile = rock::opts$get("justificationFile"),
  preventOverwriting = rock::opts$get("preventOverwriting"),
  encoding = rock::opts$get("encoding"),
  silent = rock::opts$get("silent")
)

Arguments

input
One of 1) a character string specifying the path to a file with a source; 2) an object with a loaded source as produced by a call to load_source(); 3) a character string specifying the path to a directory containing one or more sources; 4) or an object with a list of loaded sources as produced by a call to load_sources().

codes
A character vector with codes to remove.

filter
Optionally, a filter to apply to specify a subset of the source(s) to process (see get_source_filter()).

output
If specified, the recoded source(s) will be written here.

childrenReplaceParents
Whether children should be deleted (FALSE) or take their parent code's place (TRUE). This is ignored if recursiveDeletion=TRUE, in which case children are always deleted.

recursiveDeletion
Whether to also delete a code's parents (TRUE), if they have no other children, and keep doing this until the root is reached, or whether to leave parent codes alone (FALSE). This takes precedence over childrenReplaceParents.

decisionLabel
A description of the (recoding) decision that was taken.

justification
The justification for this action.

justificationFile
If specified, the justification is appended to this file. If not, it is saved to the justifier::workspace(). This can then be saved or displayed at the end of the R Markdown file or R script using justifier::save_workspace().

preventOverwriting
Whether to prevent overwriting existing files when writing the files to output.

encoding
The encoding to use.

silent
Whether to be chatty or quiet.

Value

Invisibly, the recoded source(s) or source(s) object.

Examples

### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-1.rock");

### Load example source
loadedExample <- rock::load_source(exampleFile);

### Delete two codes, moving children to the codes' parents
recoded_source <-
  rock::recode_delete(
    loadedExample,
    codes=c("childCode2", "childCode1"),
    silent=FALSE
  );

#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 110 utterances in the provided source, 110 are selected by the filter, 4 of which contain the code text.
#> --------PRE: ~specimen book. [[parentCode1]] [[childCode1]] [[intensity||2]]
#>        POST: ~specimen book. [[parentCode1>childCode2]] [[childCode1]] [[intensity||2]]
#> --UNCHANGED: ~still in their infancy. [[parentCode1>childCode1>grandchildCode2]]
#> --UNCHANGED: accompanied by English versions from the 1914 translation by H. Rackham. [[childCode1>grandchildCode2]]
#> --UNCHANGED: reasonable. The generated Lorem Ipsum is therefore always free from [[grandchildCode2]]
#> Deleted 1 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 110 utterances in the provided source, 110 are selected by the filter, 6 of which contain the code text.
#> --------PRE: Lorem Ipsum is simply dummy text of the printing and typesetting industry. [[parentCode1]]
#>        POST: Lorem Ipsum is simply dummy text of the printing and typesetting industry. [[parentCode1>childCode1]]
#> --------PRE: ~specimen book. [[parentCode1]] [[intensity||2]]
#>        POST: ~specimen book. [[parentCode1]] [[childCode1]] [[intensity||2]]
#> --------PRE: using 'Content here, content here', making it look like readable English. [[parentCode1>grandchildCode1]]
#>        POST: using 'Content here, content here', making it look like readable English. [[parentCode1>childCode1>grandchildCode1]]
#> --------PRE: ~still in their infancy. [[parentCode1>grandchildCode2]]
#>        POST: ~still in their infancy. [[parentCode1>childCode1>grandchildCode2]]
#> --------PRE: by accident, sometimes on purpose (injected humour and the like). [[parentCode1>grandchildCode3]]
#>        POST: by accident, sometimes on purpose (injected humour and the like). [[parentCode1>childCode1>grandchildCode3]]
#> --------PRE: accompanied by English versions from the 1914 translation by H. Rackham. [[grandchildCode2]]
#>        POST: accompanied by English versions from the 1914 translation by H. Rackham. [[childCode1>grandchildCode2]]
#> Deleted 6 occurrences of code 'childCode1'.
#> 

### Process an entire directory
list_of_recoded_sources <-
  rock::recode_delete(
    examplePath,
    codes=c("childCode2", "childCode1"),
    silent=FALSE
  );

#> The path to a directory was passed; loading all sources.
#> 
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/attitude.dct'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/attributes.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/behavior.dct'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/ci_example_1.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/example-1.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/intention.dct'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/longer-test.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/lorum-ipsum.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/lorum-ipsum2.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/merging-test-1-primary.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/merging-test-1-secondary.rock'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/perceived_behavioral_control.dct'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/perceived_norm.dct'.
#> Loading source from file 'C:/Users/micro/Documents/R/win-library/4.1/rock/extdata/second-test-file.rock'.
#> I just loaded 14 sources.
#> 
#> A list of loaded sources was passed; processing them sequentially.
#> 
#> Processing source 'attitude.dct'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 82 utterances in the provided source, 82 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 82 utterances in the provided source, 82 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'attributes.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 15 utterances in the provided source, 15 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 15 utterances in the provided source, 15 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'behavior.dct'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 9 utterances in the provided source, 9 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 9 utterances in the provided source, 9 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'ci_example_1.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 20 utterances in the provided source, 20 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 20 utterances in the provided source, 20 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'example-1.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 110 utterances in the provided source, 110 are selected by the filter, 4 of which contain the code text.
#> --------PRE: ~specimen book. [[parentCode1]] [[childCode1]] [[intensity||2]]
#>        POST: ~specimen book. [[parentCode1>childCode2]] [[childCode1]] [[intensity||2]]
#> --UNCHANGED: ~still in their infancy. [[parentCode1>childCode1>grandchildCode2]]
#> --UNCHANGED: accompanied by English versions from the 1914 translation by H. Rackham. [[childCode1>grandchildCode2]]
#> --UNCHANGED: reasonable. The generated Lorem Ipsum is therefore always free from [[grandchildCode2]]
#> Deleted 1 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 110 utterances in the provided source, 110 are selected by the filter, 6 of which contain the code text.
#> --------PRE: Lorem Ipsum is simply dummy text of the printing and typesetting industry. [[parentCode1]]
#>        POST: Lorem Ipsum is simply dummy text of the printing and typesetting industry. [[parentCode1>childCode1]]
#> --------PRE: ~specimen book. [[parentCode1]] [[intensity||2]]
#>        POST: ~specimen book. [[parentCode1]] [[childCode1]] [[intensity||2]]
#> --------PRE: using 'Content here, content here', making it look like readable English. [[parentCode1>grandchildCode1]]
#>        POST: using 'Content here, content here', making it look like readable English. [[parentCode1>childCode1>grandchildCode1]]
#> --------PRE: ~still in their infancy. [[parentCode1>grandchildCode2]]
#>        POST: ~still in their infancy. [[parentCode1>childCode1>grandchildCode2]]
#> --------PRE: by accident, sometimes on purpose (injected humour and the like). [[parentCode1>grandchildCode3]]
#>        POST: by accident, sometimes on purpose (injected humour and the like). [[parentCode1>childCode1>grandchildCode3]]
#> --------PRE: accompanied by English versions from the 1914 translation by H. Rackham. [[grandchildCode2]]
#>        POST: accompanied by English versions from the 1914 translation by H. Rackham. [[childCode1>grandchildCode2]]
#> Deleted 6 occurrences of code 'childCode1'.
#> 
#> Processing source 'intention.dct'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 23 utterances in the provided source, 23 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 23 utterances in the provided source, 23 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'longer-test.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 146 utterances in the provided source, 146 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 146 utterances in the provided source, 146 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'lorum-ipsum.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 16 utterances in the provided source, 16 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 16 utterances in the provided source, 16 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'lorum-ipsum2.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 16 utterances in the provided source, 16 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 16 utterances in the provided source, 16 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'merging-test-1-primary.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 24 utterances in the provided source, 24 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 24 utterances in the provided source, 24 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'merging-test-1-secondary.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 23 utterances in the provided source, 23 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 23 utterances in the provided source, 23 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'perceived_behavioral_control.dct'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 31 utterances in the provided source, 31 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 31 utterances in the provided source, 31 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'perceived_norm.dct'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 31 utterances in the provided source, 31 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 31 utterances in the provided source, 31 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#> 
#> Processing source 'second-test-file.rock'.
#> Multiple codes to remove have been specified: starting sequential removal of 2 codes.
#> 
#> Removing all occurrences of code 'childCode2'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 47 utterances in the provided source, 47 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode2'.
#> 
#> Removing all occurrences of code 'childCode1'.
#> If this code has code descendency (i.e. a child code, grandchild code, etc), the direct child code will replace the removed code.
#> If this code has code ancestry (i.e. a (partial) path to the code root is specified with a parent code, potential grandparent code, etc), the ancestry specified in that coding instance will be retained, so the utterance will remain coded with the parent code (and any ancestry of that parent code will also be retained in that coding instance).
#> Out of the 47 utterances in the provided source, 47 are selected by the filter, 0 of which contain the code text.
#> Deleted 0 occurrences of code 'childCode1'.
#>