Convenience function to get a list of all available coding schemes

Convenience function to get a list of all available coding schemes

codingSchemes_get_all()

Value

A list of all available coding schemes

Examples

rock::codingSchemes_get_all();

#> $codingScheme_peterson
#> $label
#> [1] "Peterson, Peterson & Powell"
#> 
#> $codes
#> [1] "understanding"    "retrieval"        "judgment"         "response"        
#> [5] "content_adequacy"
#> 
#> $codingInstructions
#>                                                                                                                                                                                                            understanding 
#>                                                                                                                                          "Is the item wording, terminology, and structure clear and easy to understand?" 
#>                                                                                                                                                                                                                retrieval 
#> "Has the respondent ever formed an attitude about the topic? Does the respondent have the necessary knowledge to answer the question? Are the mental calculations or long-term memory retrieval requirements too great?" 
#>                                                                                                                                                                                                                 judgment 
#>                                                                          "Is the question too sensitive to yield an honest response? Is the question relevant to the respondent? Is the answer likely to be a constant?" 
#>                                                                                                                                                                                                                 response 
#>                                                                                                 "Is the desired response available and/or accurately reflected in the response options? Are the response options clear?" 
#>                                                                                                                                                                                                         content_adequacy 
#>                                                                                                                   "Do all of the items combined adequately represent the construct? Are there items that do not belong?" 
#> 
#> $description
#> [1] ""
#> 
#> $source
#> [1] "Christina Hamme Peterson, N. Andrew Peterson & Kristen Gilmore Powell (2017) Cognitive Interviewing for Item Development: Validity Evidence Based on Content and Response Processes, Measurement and Evaluation in Counseling and Development, 50:4, 217-223, https://10.1080/07481756.2017.1339564"
#> 
#> attr(,"class")
#> [1] "rock_codingScheme"
#> 
#> $codingScheme_levine
#> $label
#> [1] "Levine, Fowler & Brown"
#> 
#> $codes
#> [1] "comprehension" "knowledge"     "inapplicable"  "construct"    
#> [5] "subtle"        "general"      
#> 
#> $codingInstructions
#>                                                                                   comprehension 
#>        "Items with unclear or ambigous terms, failed to understand the questions consistently." 
#>                                                                                       knowledge 
#>                          "Items for which respondents lacked information to answer a question." 
#>                                                                                    inapplicable 
#> "Items measuring construct that are inapplicable for many respondents (e.g. made assumptions)." 
#>                                                                                       construct 
#>                                               "Items failed to measure the intended construct." 
#>                                                                                          subtle 
#>                        "Items making discriminations that are too subtle for many respondents." 
#>                                                                                         general 
#>              "Several other general issues associated with the development of a questionnaire." 
#> 
#> $description
#> [1] ""
#> 
#> $source
#> [1] "Levine, R., Fowler, F., Jr., & Brown, J. (2005).Role of cognitive testing in the development of the CAHPS hospital survey. Health Research and Educational Trust, 40, 2037-2056. https://doi.org/10.1111/j.1475-6773.2005.00472.x"
#> 
#> attr(,"class")
#> [1] "rock_codingScheme"
#> 
#> $codingScheme_willis
#> $label
#> [1] "Willis, 1999"
#> 
#> $codes
#>  [1] "clarity"                          "clarity>wording"                 
#>  [3] "clarity>technical_term"           "clarity>vague"                   
#>  [5] "clarity>reference_periods"        "knowledge"                       
#>  [7] "knowledge>knowledge"              "knowlede>recall"                 
#>  [9] "knowledge>computation"            "assumptions"                     
#> [11] "assumptions>inappropriate"        "assumptions>constant_behavior"   
#> [13] "assumptions>double_barrelled"     "response_categories"             
#> [15] "response_categories>missing"      "response_categories>mismatch"    
#> [17] "response_categories>vague"        "response_categories>open_ended"  
#> [19] "response_categories>overlapping"  "response_categories>order"       
#> [21] "sensitivity"                      "sensitivity>general_content"     
#> [23] "sensitivity>specific_wording"     "sensitivity>social_acceptability"
#> [25] "instructions"                     "formatting"                      
#> 
#> $codingInstructions
#>                                                       clarity 
#>              "Problems with intent or meaning of a question." 
#>                                                     knowledge 
#> "Likely not to know or have trouble remembering information." 
#>                                                   assumptions 
#>              "Problems with assumptions or underlying logic." 
#>                                           response_categories 
#>                      "Problems with the response categories." 
#>                                                   sensitivity 
#>                           "Sensitive nature or wording/bias." 
#>                                                  instructions 
#> "Problems with introductions, instructions, or explanations." 
#>                                                    formatting 
#>                        "Problems with lay-out or formatting." 
#> 
#> $description
#> [1] ""
#> 
#> $source
#> [1] ""
#> 
#> attr(,"class")
#> [1] "rock_codingScheme"
#>