Create an overview of coded fragments

Collect all coded utterances and optionally add some context (utterances before and utterances after) to create ann overview of all coded fragments per code.

collect_coded_fragments(x, codes = ".*", context = 0, heading = NULL,
  headingLevel = 2, rawResult = FALSE, output = NULL,
  cleanUtterances = TRUE, silent = TRUE)

Arguments

x
The parsed source(s) as provided by rock::parse_source or rock::parse_sources.

codes
The regular expression that matches the codes to include

context
How many utterances before and after the target utterances to include in the fragments.

heading
Optionally, a title to include in the output. The title will be prefixed with headingLevel hashes (#), and the codes with headingLevel+1 hashes. If NULL (the default), a heading will be generated that includes the collected codes if those are five or less. If a character value is specified, that will be used. To omit a heading, set to anything that is not NULL or a character vector (e.g. FALSE). If no heading is used, the code prefix will be headingLevel hashes, instead of headingLevel+1 hashes.

headingLevel
The number of hashes to insert before the headings.

rawResult
Whether to return the raw result, a list of the fragments, or one character value in markdown format.

output
Here, a path and filename can be provided where the result will be written. If provided, the result will be returned invisibly.

cleanUtterances
Whether to use the clean or the raw utterances when constructing the fragments (the raw versions contain all codes).

silent
Whether to provide (FALSE) or suppress (TRUE) more detailed progress updates.

Value

Either a list of character vectors, or a single character value.

Examples

### Get path to example source
examplePath <-
  system.file("extdata", package="rock");

### Get a path to one example file
exampleFile <-
  file.path(examplePath, "example-1.rock");

### Parse single example source
parsedExample <- rock::parse_source(exampleFile);

### Show organised coded fragments in Markdown
cat(collect_coded_fragments(parsedExample));

#> ## Collected coded fragments with 0 lines of context
#> 
#> ### childCode1
#> 
#> -----
#> 
#> Lorem Ipsum is simply dummy text of the printing and typesetting industry.
#> 
#> -----
#> 
#> specimen book.
#> 
#> -----
#> 
#> ### childCode2
#> 
#> -----
#> 
#> specimen book.
#> 
#> -----
#> 
#> ### childCode3
#> 
#> -----
#> 
#> electronic typesetting, remaining essentially unchanged.
#> 
#> -----
#> 
#> ### childCode4
#> 
#> -----
#> 
#> including versions of Lorem Ipsum.
#> 
#> -----
#> 
#> ### childCode5
#> 
#> -----
#> 
#> content of a page when looking at its layout.
#> 
#> -----
#> 
#> ### grandchildCode1
#> 
#> -----
#> 
#> using 'Content here, content here', making it look like readable English.
#> 
#> -----
#> 
#> ### grandchildCode2
#> 
#> -----
#> 
#> still in their infancy.
#> 
#> -----
#> 
#> accompanied by English versions from the 1914 translation by H. Rackham.
#> 
#> -----
#> 
#> reasonable. The generated Lorem Ipsum is therefore always free from
#> 
#> -----
#> 
#> ### grandchildCode3
#> 
#> -----
#> 
#> by accident, sometimes on purpose (injected humour and the like).
#> 
#> -----
#> 
#> ### grandchildCode4
#> 
#> -----
#> 
#> Where does it come from?
#> 
#> -----
#> 
#> ### grandchildCode5
#> 
#> -----
#> 
#> old.
#> 
#> -----
#> 
#> ### grandchildCode6
#> 
#> -----
#> 
#> discovered the undoubtable source.
#> 
#> -----
#> 
#> ### grandchildCode7
#> 
#> -----
#> 
#> repetition, injected humour, or non-characteristic words etc.
#> 
#> -----
#> 
#> ### someOtherCode
#> 
#> -----
#> 
#> electronic typesetting, remaining essentially unchanged.
#> 
#> -----

### Only for the codes containing 'Code2'
cat(collect_coded_fragments(parsedExample,
                            'Code2'));

#> ## Collected coded fragments for codes 'childCode2' & 'grandchildCode2' with 0 lines of context
#> 
#> ### childCode2
#> 
#> -----
#> 
#> specimen book.
#> 
#> -----
#> 
#> ### grandchildCode2
#> 
#> -----
#> 
#> still in their infancy.
#> 
#> -----
#> 
#> accompanied by English versions from the 1914 translation by H. Rackham.
#> 
#> -----
#> 
#> reasonable. The generated Lorem Ipsum is therefore always free from
#> 
#> -----